+++ src/screens/chat/ChatRoom.js
@@ -1,30 +1,24 @@
+import {useCallback} from 'react';
 import {useSelector} from 'react-redux';
-import {
-  View,
-  StyleSheet,
-  Platform,
-  Image,
-  Text,
-  TouchableOpacity,
-} from 'react-native';
-import {GiftedChat, InputToolbar, Send} from 'react-native-gifted-chat';
-import React, {useState, useEffect} from 'react';
+import Colors from '../../utils/Colors';
 import {getAllmessages} from './ChatUtils';
+import {normalize} from '../../utils/Dimensions';
+import React, {useState, useEffect} from 'react';
+import LocalImages from '../../utils/LocalImages';
+import {showToast} from '../../utils/CommonFunctions';
+import {useNavigation} from '@react-navigation/native';
 import firestore from '@react-native-firebase/firestore';
 import {getStatusBarHeight} from 'react-native-status-bar-height';
-import LocalImages from '../../utils/LocalImages';
-import {normalize} from '../../utils/Dimensions';
-import Colors from '../../utils/Colors';
+import {View, StyleSheet, Platform, Image, Text} from 'react-native';
+import {GiftedChat, InputToolbar, Send} from 'react-native-gifted-chat';
 import ChatRoomHeader from '../../components/chatRoomHeader/ChatRoomHeader';
-import {useNavigation} from '@react-navigation/native';
-import {useCallback} from 'react';
-import {showToast} from '../../utils/CommonFunctions';
 
 export default function ChatRoom({route}) {
   const navigation = useNavigation();
-  const [messages, setMessages] = useState([]);
-  const {userId} = useSelector(store => store.userDetailsReducer);
   const {userID, image, name} = route?.params;
+  const [messages, setMessages] = useState([]);
+  const [isTyping, setisTyping] = useState(false);
+  const {userId} = useSelector(store => store.userDetailsReducer);
 
   let docid = userId > userID ? userId + '-' + userID : userID + '-' + userId;
 
@@ -58,6 +52,7 @@ export default function ChatRoom({route}) {
       .collection('messages')
       .add({...mymsg});
   };
+
   const handleBack = useCallback(() => {
     navigation.goBack();
   }, [navigation]);
@@ -68,6 +63,38 @@ export default function ChatRoom({route}) {
         containerStyle={styles.inputToolbarContainerStyle}
         {...props}
       />
+    );
+  };
+  const debounce = (fun, timeout) => {
+    let timer;
+    return args => {
+      clearTimeout(timer);
+      timer = setTimeout(() => {
+        fun(args);
+      }, timeout);
+      setisTyping(true);
+    };
+  };
+
+  const startTyping = debounce(() => {
+    setisTyping(false);
+  }, 1000);
+
+  const detectTyping = text => {
+    if (text.length > 0) startTyping(false);
+  };
+
+  const renderFooter = () => {
+    return (
+      <View
+        style={{
+          marginBottom: 20,
+          height: 20,
+          backgroundColor: 'green',
+        }}>
+        <Text style={{color: 'red'}}>{`${isTyping}`}</Text>
+      </View>
+      // else return null;
     );
   };
 
@@ -109,7 +136,8 @@ export default function ChatRoom({route}) {
         }}
         renderInputToolbar={renderInputToolbar}
         renderSend={renderSend}
-        isTyping={true}
+        onInputTextChanged={detectTyping}
+        renderFooter={renderFooter}
       />
     </View>
   );
