+++ src/screens/home/Chats.js
@@ -2,7 +2,7 @@ import Colors from '../../utils/Colors';
 import {useDispatch, useSelector} from 'react-redux';
 import React, {useState, useEffect} from 'react';
 import {FlatList, StyleSheet, View} from 'react-native';
-import {getDatafromFirebase, showToast} from '../../utils/CommonFunctions';
+import {showToast} from '../../utils/CommonFunctions';
 import ChatListRender from '../../components/chatListRender/ChatListRender';
 import {requestDataAllUsers} from '../../redux/userDetails/action';
 
@@ -10,7 +10,7 @@ const Chats = () => {
   const [users, setAllUsers] = useState();
   const dispatch = useDispatch();
   const {userId} = useSelector(store => store.userDetailsReducer);
-
+  console.log('wertyuiop;lkjhgfds', users);
   const getAllUsers = () => {
     dispatch(
       requestDataAllUsers(
@@ -32,10 +32,10 @@ const Chats = () => {
   const onRender = ({item}) => {
     return (
       <ChatListRender
-        name={item.name}
-        chatImage={item.image}
         id={item.id}
+        name={item.name}
         status={item.status}
+        chatImage={item.image}
       />
     );
   };
@@ -50,6 +50,9 @@ const Chats = () => {
         data={users}
         renderItem={onRender}
         ItemSeparatorComponent={flatListItemSeparator}
+        keyExtractor={item => {
+          item.id;
+        }}
       />
     </View>
   );
