--- src/screens/home/HomeUtils.js
@@ -1,1 +1,0 @@
-
